third_party/blink/web_tests/flag-specific/disable-layout-ng/external/wpt/css/css-tables/bounding-box-computation-2-expected.txt

This is a testharness.js-based test.
PASS Control test: Table width is 120px
PASS Control test: Table height is 120px
PASS Table-cell is 100px wide
PASS Table-cell is 100px tall
PASS Table-row is 100px wide
PASS Table-row is 100px tall
PASS Table-row-group is 100px wide
PASS Table-row-group is 100px tall
FAIL Table-column is 100px wide assert_equals: expected 100 but got 0
FAIL Table-column is 100px tall assert_equals: expected 100 but got 0
FAIL Table-column-group is 100px wide assert_equals: expected 100 but got 0
FAIL Table-column-group is 100px tall assert_equals: expected 100 but got 0
Harness: the test ran to completion.
